+++ service/src/protocol/error_code.rs
@@ -87,6 +87,8 @@ pub enum ErrorCode {
     ServiceAlreadyRunning = 4096 + 10,
     /// Store directory is not specified
     StoreDirUnspecified = 4096 + 11,
+    /// Mount directory is not specified
+    MountDirUnspecified = 4096 + 12,
 
     /// Unspecified error
     Other = 65535,
@@ -109,12 +111,12 @@ impl ToErrorCode for Error {
             Self::NotFound => ErrorCode::NotFound,
             Self::Ambiguous => ErrorCode::Ambiguous,
             Self::Busy => ErrorCode::ResourceBusy,
-            Self::OperationNotSupported => ErrorCode::Unsupported,
             Self::OperationInterrupted => ErrorCode::Interrupted,
             Self::PermissionDenied => ErrorCode::PermissionDenied,
             Self::Repository(error) => error.to_error_code(),
             Self::Store(_) => ErrorCode::StoreError,
             Self::StoreDirUnspecified => ErrorCode::StoreDirUnspecified,
+            Self::MountDirUnspecified => ErrorCode::MountDirUnspecified,
             Self::TlsCertificatesNotFound => ErrorCode::TlsCertificatesNotFound,
             Self::TlsCertificatesInvalid(_) => ErrorCode::TlsCertificatesInvalid,
             Self::TlsConfig(_) => ErrorCode::TlsConfigError,
@@ -149,7 +151,7 @@ impl ToErrorCode for ClientError {
             Self::Connect(_) => ErrorCode::ConnectionRefused,
             Self::Disconnected => ErrorCode::ConnectionAborted,
             Self::InvalidArgument => ErrorCode::InvalidInput,
-            Self::InvalidEndpoint => ErrorCode::InvalidInput,
+            Self::InvalidEndpoint(_) => ErrorCode::InvalidInput,
             Self::Io(_) => ErrorCode::Other,
             Self::Read(ReadError::Receive(_)) => ErrorCode::TransportError,
             Self::Read(ReadError::Decode(_)) => ErrorCode::InvalidData,