+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased](https://github.com/equalitie/ouisync/compare/v0.9.1...master) -->
+## [Unreleased](https://github.com/equalitie/ouisync/compare/v0.9.1...master)
+
+- Improve some error messages in the CLI app.
 
 ## [v0.9.1](https://github.com/equalitie/ouisync/compare/v0.9.0...v0.9.1) - 2025-06-26